+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-07-06T10:12:53.829Z
+Generated: 2026-07-07T09:35:41.014Z
 
 - 2026-07-05 • Microsoft Defender for Cloud • Deprecation • Discovery and posture for serverless container workloads is now generally available
 - 2026-07-01 • Microsoft Defender for Cloud • Deprecation • Plan enablement API now blocks onboarding to five deprecated Defender plans
@@ -13,6 +13,7 @@ Generated: 2026-07-06T10:12:53.829Z
 - 2026-07-01 • Microsoft Sentinel • Action required • [Updated] Call to action: update automation by July 1, 2026 - Account Name is now consistently the UPN prefix for analytics rule alerts
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
+- 2026-07-01 • Microsoft Sentinel • Preview • Custom detection rules support in Microsoft Sentinel Repositories (Preview)
 - 2026-07-01 • Unified SecOps • Deprecation • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
 - 2026-06-30 • Microsoft Defender for Cloud • GA • Expanded multicloud security coverage is now generally available
 - 2026-06-30 • Microsoft Defender for Cloud • GA • General availability of Defender for Key Vault in Azure Government cloud
@@ -21,4 +22,3 @@ Generated: 2026-07-06T10:12:53.829Z
 - 2026-06-17 • Microsoft Defender for Cloud • GA • API security posture management for Function Apps and Logic Apps is now generally available
 - 2026-06-15 • Microsoft Sentinel • Action required • Call to action: update older Microsoft Sentinel content as code (Sentinel repositories) API versions before June 15, 2026
 - 2026-06-08 • Microsoft Defender for Cloud • Preview • New multicloud security recommendations now in public preview
-- 2026-06-04 • Microsoft Defender for Cloud • GA • SQL Vulnerability Assessment Express Configuration is now generally available for Azure SQL Managed Instance and Azure Synapse Analytics workspaces